Make ModelarDBFlightClient into a clean wrapper for FlightClient base endpoints
